+++ gpaw/grid_descriptor.py
@@ -450,9 +450,11 @@ class GridDescriptor(Domain):
             B_g = np.zeros_like(A_g)
             for s, op_cc in enumerate(op_scc):
                 if ft_sc is None:
-                    _gpaw.symmetrize(A_g, B_g, op_cc)
+                    _gpaw.symmetrize(A_g, B_g, op_cc, 1 - self.pbc_c)
                 else:
-                    _gpaw.symmetrize_ft(A_g, B_g, op_cc, ft_sc[s])
+                    t_c = (ft_sc[s] * self.N_c).round().astype(int)
+                    _gpaw.symmetrize_ft(A_g, B_g, op_cc, t_c,
+                                        1 - self.pbc_c)
         else:
             B_g = None
         self.distribute(B_g, a_g)